git-svn-id: svn://123.57.173.36/meijia-parent@2139 52ac4245-7803-4e87-b018-04b7c5ce731e

+++ simi-dao/src/main/java/com/simi/po/model/total/TotalHit.java
@@ -7,7 +7,7 @@ public class TotalHit {
 
     private Long linkId;
 
-    private Long total;
+    private int total;
 
     private Long addTime;
 
@@ -35,11 +35,11 @@ public class TotalHit {
         this.linkId = linkId;
     }
 
-    public Long getTotal() {
+    public int getTotal() {
         return total;
     }
 
-    public void setTotal(Long total) {
+    public void setTotal(int total) {
         this.total = total;
     }
